Move context sessionInfo methods to managerutil package
