+++ cmd/traffic/cmd/manager/managerutil/ctx.go
@@ -1,4 +1,4 @@
-package manager
+package managerutil
 
 import (
 	"context"
